+++ packages/grafana-data/src/transformations/index.ts
@@ -1,8 +1,7 @@
 export * from './matchers/ids';
 export * from './transformers/ids';
-export * from './transformersRegistry';
-export * from './matchers/matchers';
-export * from './transformersRegistry';
-export * from './transformDataFrame';
+export * from './matchers';
+export * from './transformers';
+export * from './fieldReducer';
 export { FilterFieldsByNameTransformerOptions } from './transformers/filterByName';
 export { ReduceTransformerOptions } from './transformers/reduce';
