+++ packages/grafana-data/src/transformations/matchers/fieldTypeMatcher.test.ts
@@ -1,5 +1,5 @@
 import { FieldType } from '../../types/dataFrame';
-import { fieldMatchers } from './matchers';
+import { fieldMatchers } from '../matchers';
 import { FieldMatcherID } from './ids';
 import { toDataFrame } from '../../dataframe/processDataFrame';
