+++ packages/grafana-data/src/transformations/matchers/matchers.test.ts
@@ -1,4 +1,4 @@
-import { fieldMatchers } from './matchers';
+import { fieldMatchers } from '../matchers';
 import { FieldMatcherID } from './ids';
 
 describe('Matchers', () => {
