+++ packages/grafana-data/src/transformations/matchers/nameMatcher.test.ts
@@ -1,4 +1,4 @@
-import { getFieldMatcher } from './matchers';
+import { getFieldMatcher } from '../matchers';
 import { FieldMatcherID } from './ids';
 import { toDataFrame } from '../../dataframe/processDataFrame';
